+++ src/core/include/engine/renderingUtils.hpp
@@ -1,6 +1,7 @@
 #pragma once
 #include "smoothImplicit.hpp"
 #include "macros.hpp"
+#include "shaderTypes.hpp"
 
 #include <array>
 #include <map>
@@ -16,19 +17,6 @@
 
 
 
-enum GLSLType {
-	FLOAT,
-	INT,
-	VEC2,
-	VEC3,
-	VEC4,
-	MAT2,
-	MAT3,
-	MAT4,
-	SAMPLER1D,
-	SAMPLER2D,
-	SAMPLER3D
-};
 
 class Texture {
 public:
